Remove unused static_cast

+++ examples/chef/common/clusters/channel/ChannelManager.h
@@ -73,6 +73,6 @@ protected:
 private:
     bool isChannelMatched(const ChannelInfoType & channel, const chip::CharSpan & match);
     static constexpr uint32_t mDynamicEndpointFeatureMap =
-        static_cast<uint32_t>(chip::BitMask<Feature, uint32_t>(Feature::kChannelList, Feature::kLineupInfo).Raw());
+        chip::BitMask<Feature, uint32_t>(Feature::kChannelList, Feature::kLineupInfo).Raw();
     static constexpr uint16_t kClusterRevision = 2;
 };
